refactor: FirstMenu 폴더 생성 및 파일 분리
